pyproject.toml
[project]
name = "hed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "google-genai>=1.25.0",
    "hedtools>=0.5.0",
    "ipykernel>=6.29.5",
    "jinja2>=3.1.6",
    "ollama>=0.5.1",
    "python-dotenv>=1.1.1",
]
